Add FileKeyProvider.
This provider reads keys from the path set in ELASTICRYPT_FILE_KEYRING
environment variable. If the variable is not set an attempt is made to
read keys from CONF_DIR/keyring.

If the keyring path is a directory it must contain a key file for each
encrypted index (file names must match index names). If the keyring path
points to a file the same key will be used for all indices.

The provider ensures that key files are read exactly once at startup.